requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
black==23.3.0
certifi==2023.5.7
cffi==1.15.0
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
contourpy==1.0.5
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
dicom2nifti==2.4.8
docopt==0.6.2
duckdb==0.8.0
entrypoints==0.4
enum34==1.1.10
executing==1.2.0
filelock==3.12.0
fonttools==4.25.0
frozenlist==1.3.3
fsspec==2023.5.0
future==0.18.3
idna==3.4
imageio==2.30.0
imgaug==0.4.0
ipykernel==6.15.0
ipython==8.13.2
isort==5.12.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter-client==7.3.4
jupyter_core==5.3.0
kiwisolver==1.4.4
lazy_loader==0.2
lightning-utilities==0.8.0
lit==16.0.5
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
monai==1.1.0
mpmath==1.3.0
multidict==6.0.4
munkres==1.1.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
nibabel==5.1.0
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
pandas-ml==0.6.1
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.0.1
pipreqs==0.4.13
platformdirs==3.5.1
ply==3.11
prompt-toolkit==3.0.38
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pycparser==2.21
pydicom==2.3.1
Pygments==2.15.1
pylance==0.4.14
pyparsing==3.0.9
PyQt5-sip==12.11.0
python-dateutil==2.8.2
python-gdcm==3.0.22
pytorch-lightning==2.0.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
requests==2.31.0
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
setuptools==66.0.0
shapely==2.0.1
SimpleITK==2.2.1
sip==6.6.2
six==1.16.0
stack-data==0.6.2
sympy==1.12
threadpoolctl==3.1.0
tifffile==2023.4.12
tokenize-rt==5.0.0
toml==0.10.2
tomli==2.0.1
torch==2.0.1
torchmetrics==0.11.4
torchvision==0.15.2
tornado==6.1
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.6.1
tzdata==2023.3
urllib3==2.0.2
wcwidth==0.2.6
wheel==0.38.4
wordcloud==1.9.2
yarg==0.1.9
yarl==1.9.2
